攻撃モーションの作り込み(踏み込み・コンボ・ダッシュ攻撃)と移動のブレンドツリー
移動
- 移動の8方向ブレンドツリーを追加(LocomotionBlendSpace.h)。向き(前後左右)×歩き/走りを速度で混ぜ、足運びの位相を共有する
- 足運びを実際の移動速度に合わせて再生速度を決め、足が地面を滑らないようにした
- 歩きを20、ダッシュを45へ下げ、キャラクターの体格に合わせた
- ロックオンしても体は敵を向かない(モンハンと同じくカメラだけが敵を追う)
- ゲームのウィンドウが前面にないときにキー入力を拾っていた不具合を修正

攻撃
- 攻撃中は脚と腰もクリップから取り、腰の水平移動をキャラクターの位置へ移す踏み込み(ルートモーション)を実装
- 攻撃の予兆・判定・硬直を、コンボの段ごとにクリップの秒で持つ表から決める(見た目の振りと判定を一致させる)
- 弱コンボを 振り下ろし → 斬り上げ → 横薙ぎ に組み直し、段ごとに剣の向きを変えて見分けられるようにした
- 走りながら弱攻撃でダッシュ攻撃(跳躍斬り)が出る
- 強攻撃の2段目が振り切る前に3段目へ移っていたのを修正
- しゃがんで始まるクリップで足が浮く問題を修正(腰の高さをクリップの値で合わせる)
- 踏み込みが後ろ向きだった不具合を修正(読み込み時のZ反転で前が-Zになっていた)
- 攻撃中に体が斜めを向く不具合を修正(腰の回転の基準を待機の構えからバインド姿勢へ)
- 敵がいなくても剣を振れるようにした(デバッグ表示の「敵を全員消す(素振り)」/ dev_settings.ini の no_enemy=1)

その他
- F3のゲーム内モーションエディターを外した(デバッグツールの重複を避けるため)
- アニメーションは経過時間で進める(敵のアニメーション送りと振り向きがfps依存だったのを修正)
- 実装ログ・Codex引き継ぎメモ・技術解説を更新

+++ scene/GameScene.h
@@ -82,7 +82,8 @@ private:
 	aiAnimation* m_enemyAttackAnimation = nullptr;
 	aiAnimation* m_enemyDieAnimation = nullptr;
 	int m_enemyAnimationFrame = 0;
-	int m_enemyAnimationTick = 0;
+	// 次のキーまでの進み具合(0〜1)。経過時間で進める(fpsに依存させない)。
+	float m_enemyAnimationTick = 0.0f;
 	// カットシーン終了時のIdle姿勢を、戦闘開始時のWalk先頭へ短時間でつなぐ。
 	static constexpr float ENEMY_INTRO_BATTLE_BLEND_SECONDS = 0.18f;
 	bool m_enemyIntroBattleBlendActive = false;
